Remove utils/path module
