+++ exonum-java-binding/core/rust/src/utils/mod.rs
@@ -21,7 +21,6 @@ mod errors;
 mod handle;
 mod jni;
 pub mod jni_cache;
-mod path;
 pub mod services;
 mod time_service;
 
@@ -32,4 +31,3 @@ pub use self::errors::{
 };
 pub use self::handle::{as_handle, cast_handle, drop_handle, to_handle, Handle};
 pub use self::jni::{get_class_name, get_exception_message};
-pub use self::path::executable_directory;
